authorization added

+++ ToDo.BL/ToDoService.cs
@@ -12,6 +12,8 @@ using System.Reflection.Metadata.Ecma335;
 using Common.BL;
 using Common.Repository;
 using ToDo.BL;
+using Common.Api.Service;
+using Common.BL.Exeptions;
 
 namespace ToDoBL
 {
@@ -19,27 +21,46 @@ namespace ToDoBL
     {
         private readonly IRepository<TodoItem> _toDoRepository;
         private readonly IRepository<AppUser> _users;
+        private readonly ICurrentUserService _currentUserService;
         private readonly IMapper _mapper;
+        private int _currentUserId;
+        private List<string> _userRoles;
         public ToDoService(
             IRepository<TodoItem> repository,
+            ICurrentUserService currentUserService,
             IRepository<AppUser> user,
             IMapper mapper)
         {
             _mapper = mapper;
 
+            _currentUserService = currentUserService;
             _toDoRepository = repository;
-
             _users = user;
-            
+            _currentUserId = int.Parse(_currentUserService.CurrentUserId);
+            _userRoles = _currentUserService.CurrentUserRoles.ToList();
         }
         public async Task<IEnumerable<TodoItem>> GetListAsync(int? offset, int? ownerId, string? lable, int? limit = 10, CancellationToken cancellationToken = default)
         {
-            return await _toDoRepository.GetListAsync(
+
+            if(_userRoles.Contains("Admin"))
+            {
+                return await _toDoRepository.GetListAsync(
                 offset,
                 limit,
                 d => (string.IsNullOrWhiteSpace(lable) || d.Label.Contains(lable, StringComparison.InvariantCultureIgnoreCase))
                 && (ownerId == null || d.OwnerId == ownerId.Value),
-                t => t.Id); ;
+                t => t.Id);
+            }
+            else
+            {
+                return await _toDoRepository.GetListAsync(
+                offset,
+                limit,
+                d => (string.IsNullOrWhiteSpace(lable) || d.Label.Contains(lable, StringComparison.InvariantCultureIgnoreCase))
+                && (d.OwnerId == _currentUserId),
+                t => t.Id);
+            }
+            
         }
 
         public async Task<TodoItem> GetByIdAsync(int id, CancellationToken cancellationToken)
@@ -49,7 +70,14 @@ namespace ToDoBL
             {
                 throw new NotFoundExeption(new {Id = id});
             }
-            return item;
+            else if(CheckAccess(item.OwnerId))
+            {
+                return item;
+            }
+            else
+            {
+                throw new ForbidenExeption("Access denied");
+            }
         }
         public async Task<TodoItem> GetByIdIsDoneAsync(int id, CancellationToken cancellationToken)
         {
@@ -58,18 +86,40 @@ namespace ToDoBL
             {
                 throw new NotFoundExeption(new { Id = id});
             }
-            return result;
+            if (CheckAccess(result.OwnerId))
+            {
+                return result;
+            }
+            else throw new ForbidenExeption("Access denied");
         }
 
         public async Task<bool> DeleteAsync(int id, CancellationToken cancellationToken)
         {
+
             var item = await GetByIdAsync(id, cancellationToken);
-            return await _toDoRepository.DeleteAsync(item);
+            if(CheckAccess(item.OwnerId))
+            {
+                return await _toDoRepository.DeleteAsync(item);
+            }
+            else
+            {
+                throw new ForbidenExeption("Access denied");
+            }
+            
         }
 
         public async Task<TodoItem> AddAsync(CreateToDo item, CancellationToken cancellationToken)
         {
-            int ownerId = item.OwnerId;
+            int ownerId;
+            if (_userRoles.Contains("Admin"))
+            {
+                ownerId = item.OwnerId;
+            }
+            else
+            {
+                ownerId = _currentUserId;
+                item.OwnerId = _currentUserId;
+            }
             var user = await _users.SingleOrDefaultAsync(u => u.Id == ownerId);
 
             if(user == null)
@@ -105,34 +155,50 @@ namespace ToDoBL
             todoEntity.UpdatedDate = DateTime.UtcNow;
             todoEntity.User = user;
 
-            var updatedItem = await _toDoRepository.UpdateAsync(todoEntity, cancellationToken);
-
-            if(updatedItem is null)
+            if(CheckAccess(todoEntity.OwnerId))
             {
-                throw new Exception("Can not update ToDo item");
+                var updatedItem = await _toDoRepository.UpdateAsync(todoEntity, cancellationToken);
+                if (updatedItem is null)
+                {
+                    throw new Exception("Can not update ToDo item");
+                }
+                return updatedItem;
             }
-
-            return updatedItem;
+            else
+            {
+                throw new ForbidenExeption("Access denied");
+            }
         }
 
         public async Task<TodoItem> PutchAsync(int id, bool isDone, CancellationToken cancellationToken)
         {
-            var TodoItem = await GetByIdAsync(id, cancellationToken);
-            if(TodoItem == null)
+            var todoItem = await GetByIdAsync(id, cancellationToken);
+            if(todoItem == null)
             {
                 throw new NotFoundExeption("Todo item not found");
             }
 
-            TodoItem.IsDone = isDone;
+            todoItem.IsDone = isDone;
 
-            var putchedItem = await _toDoRepository.UpdateAsync(TodoItem);
-
-            if(putchedItem is null)
+            if(CheckAccess(todoItem.OwnerId))
             {
-                throw new BadRequestExeption("Can not create item");
-            }
 
-            return putchedItem;
+                var putchedItem = await _toDoRepository.UpdateAsync(todoItem);
+                if (putchedItem is null)
+                {
+                    throw new BadRequestExeption("Can not create item");
+                }
+                return putchedItem;
+            }
+            else
+            {
+                throw new ForbidenExeption("Access denied");
+            }
+        }
+
+        private bool CheckAccess(int id)
+        {
+            return _currentUserId == id || _userRoles.Contains("Admin");
         }
 
     }
